Convert first files to TypeScript
